Use grains to track Connections, Users, Groups and the Hubs
Tracking connections and users/groups via their own grains helps to fan out the workload and state size and reduce bottlenecks on central grains. This also allows us to locate users and connections directly, taking full advantage of the Orlean grains directory.

The hub lifetime manager is an observer of all signalr grains relevant to its own silo. This allows it to be notified of any requirement to push out a message, regardless of which silo the message originates from or which silo the observed signalr grains are on.

Cleanup of persisted grain state (the observers for the grain) is done via a reminder with a configurable period, which is defaulted to 1 day.