Add sample client and server
The server starts a single silo, with grains and reminders backed by memory persistence. It has a single hub which on receiving a message from the client, adds the client to Group g1, will reply back with the same message to the Caller, All Clients, and Group g1.

The client connects to the server and allows message entry via the command line. There is an optional integer argument to the client to define how many connections are created, useful for exercising the SendToAll and SendToGroup options, as well as a little load testing. If no argument is passed, a single connection is created.